style: Integrate clang-tidy and apply code modernization
- Added .clang-tidy configuration with a focus on bugprone, performance, and modernization checks.
- Implemented the Rule of Five by explicitly deleting copy/move operations for App and Window classes.
- Updated logging to use C++23 std::print and improved thread-safe exit handling in assertions.
- Applied various code quality improvements, including const-correctness, static_cast usage, and string_view optimizations.

+++ LustraLib/include/LustraLib/Logger.h
@@ -20,7 +20,7 @@
 
 namespace LustraLib
 {
-	enum OutputLevel : uint32_t
+	enum OutputLevel : uint8_t
 	{
 		OutputLevelTrace = 0u,
 		OutputLevelDebug,
